Use byte as class id type for horizontal class merging
Bug: b/528262429

+++ src/test/java8/classmerging/com/android/tools/r8/classmerging/horizontal/ClInitMergeSuperTypeApiLevelTest.java
@@ -85,7 +85,7 @@ public class ClInitMergeSuperTypeApiLevelTest extends TestBase {
                   canUseExecutable() ? 1 : 2,
                   clazz.allMethods(MethodSubject::isInstanceInitializer).size());
               if (canUseExecutable()) {
-                MethodSubject constructorInit = clazz.init(Executable.class.getTypeName(), "int");
+                MethodSubject constructorInit = clazz.init(Executable.class.getTypeName(), "byte");
                 assertThat(
                     constructorInit,
                     isAbsentIf(parameters.canUseJavaLangInvokeVarHandleStoreStoreFence()));
